requirements.txt
python-telegram-bot>=21.9
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
websockets>=14.2
google-genai>=1.14.0
anthropic>=0.45.0
python-dotenv>=1.0.1
Pillow>=11.1.0
pydantic>=2.10.0
